🐛Fix: Telegram modal api
