+++ src/common/FileUploadComponent.tsx
@@ -3,13 +3,14 @@ import { Box, Button, Flex, Input, Text } from '@chakra-ui/react';
 import { ModalContext } from '../context/ModalContext';
 import useChatStore from '../state/chat';
 import { ChatMessage } from '../state/chat';
+import { dmTelegramMembers } from '../api/dmtelegram';
 
 interface FileUploadComponentProps {
   onFileSelect: (file: File) => void;
 }
 
  const FileUploadComponent: React.FC<FileUploadComponentProps> = ({ onFileSelect }) => {
-  const { file, setFile, userPass } = useContext(ModalContext);
+  const { file, setFile, userPass} = useContext(ModalContext);
   // const [selectedFile, setSelectedFile] = useState<File | null>(null);
   const {setShowFileUploadModal, addMessage, currentFunctionArguments} = useChatStore((state) => state)
 
@@ -20,6 +21,18 @@ interface FileUploadComponentProps {
        //onFileSelect(file);
       const selectedFile = e.target.files[0];
       setFile(selectedFile);
+      const apikey = localStorage.getItem('telegramApiKey') || '';
+      const apihash = localStorage.getItem('telegramApiHash') || '';
+      const phone = localStorage.getItem('telegramPhoneNumber') || '';
+      if(!selectedFile) throw Error("No file passed!")
+      const res = await dmTelegramMembers(apikey, apihash, phone, currentFunctionArguments.msg,selectedFile)
+      const newMessage: ChatMessage = {
+        id: Date.now(),
+        sender: 'AI assistant',
+        content: "Messages Send Successfully",
+        timestamp: Date.now(),
+      };
+      addMessage(newMessage);
       setShowFileUploadModal(false)
     }
   };
@@ -42,7 +55,7 @@ interface FileUploadComponentProps {
         type="file"
         display="none"
         onChange={handleFileChange}
-        accept=".pdf,.doc,.docx,.jpg,.png"
+        accept=".pdf,.doc,.docx,.jpg,.png,.csv"
       />
     </Flex>
   );
